+++ config.el
@@ -107,6 +107,18 @@ END
 	   (let ((end (point)))
 	     (buffer-substring-no-properties bgn end)))))) lex)))
 
+(defun revert-all-buffers ()
+  "Refreshes all open buffers from their respective files."
+  (interactive)
+  (dolist (buf (buffer-list))
+    (with-current-buffer buf
+      (when (and
+	     (buffer-file-name)
+	     (file-exists-p (buffer-file-name))
+	     (not (buffer-modified-p)))
+        (revert-buffer t t t))))
+  (message "Refreshed open files."))
+
 (if (eab/ondaemon (eab/server-C))
     (progn
       (setq server-port 5001)